refactor: migrate unsub digest
