--- serverless/unsubscribe-digest/.env-example
@@ -1,13 +1,0 @@
-# Required
-DB_URI="postgres://localhost:5432/postmangovsg_dev"
-DB_READ_REPLICA_URI="postgres://localhost:5432/postmangovsg_dev"
-UNSUBSCRIBE_GUIDE_URL="https://guide.postman.gov.sg"
-# SENTRY_DSN="https://123.ingest.sentry.io/456"
-# CRONITOR_CODE="ABCDE"
-
-# You can override other config too
-# See config.ts for the full list
-# SES_HOST=""
-# SES_PASS=""
-# SES_USER=""
-# SES_PORT=""
